Dynamics SDK for React Native v12.0
  - added support for BlackBerry Dynamics SDK for iOS v12.0, Launcher v12
  - added support for BlackBerry Dynamics SDK for Android v12.0, Launcher v12
  - added support for React Native 0.70, 0.71, 0.72 versions
  - deprecated support for React Native 0.66 - 0.69 versions
  - sample apps are updated to v0.70.0 RN
  - implemented secure file download in WebView UI component
  - upgraded core for forked modules to use 0.70 as base version: Clipboard, Networking, AsyncStorage, SQLite, FileSystem,
  - upgraded core for forked ui-components to use 0.70 as base version: WebView, Text, TextInput
  - updated steps how to integrate PlayIntegrity feature
  - added fix for issue: RN BBD Android 'Edit File Service' not receiving event AppKinetics when Annotate Service is enabled
  - added other minor improvements and defect fixes

+++ ui-components/BlackBerry-Dynamics-for-React-Native-Text/scripts/add_bbd_text_widget.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 /**
- * Copyright (c) 2022 BlackBerry Limited. All Rights Reserved.
+ * Copyright (c) 2023 BlackBerry Limited. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -70,12 +70,12 @@
   }
 
   function addBbdTextWidget (filePath, widget, insertBefore) {
-    var fileContent = fs.readFileSync(filePath, 'utf-8');
-
-    if (fileContent.indexOf(bbdTextWidgetName) >= 0) {return;}
-
-    fileContent = fileContent.replace(insertBefore, widget + insertBefore);
-    fs.writeFileSync(filePath, fileContent, 'utf-8');
+    if (fs.existsSync(filePath)){
+      var fileContent = fs.readFileSync(filePath, 'utf-8');
+      if (fileContent.indexOf(bbdTextWidgetName) >= 0) {return;}
+      fileContent = fileContent.replace(insertBefore, widget + insertBefore);
+      fs.writeFileSync(filePath, fileContent, 'utf-8');
   }
+}
 
 })();
